refactor: extract shared combo-box scroller logic to mixin
