+++ packages/combo-box/src/vaadin-combo-box-scroller.d.ts
@@ -1,0 +1,19 @@
+/**
+ * @license
+ * Copyright (c) 2015 - 2023 Vaadin Ltd.
+ * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
+ */
+import { ComboBoxScrollerMixin } from './vaadin-combo-box-overlay-mixin.js';
+
+/**
+ * An element used internally by `<vaadin-combo-box>`. Not intended to be used separately.
+ */
+declare class ComboBoxScroller extends ComboBoxScrollerMixin(Overlay) {}
+
+declare global {
+  interface HTMLElementTagNameMap {
+    'vaadin-combo-box-scroller': ComboBoxScroller;
+  }
+}
+
+export { ComboBoxScroller };
